+++ RELEASE_NOTES.md
@@ -1,5 +1,5 @@
 ### New Features
 
 - Implement the missing _create_ and _update_ endpoints for Brands.
-
+- Implement [Channels API](https://developer.bigcommerce.com/api-reference/store-management/channels)